requirements.txt
beautifulsoup4==4.6.0
matplotlib==2.1.2
nltk==3.2.5
numpy==1.13.3
python-dateutil==2.6.1
python-Levenshtein==0.12.0
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
sklearn==0.0
urllib3==1.22
